+++ MAINTAINER.md
@@ -17,11 +17,6 @@ Maintainer guide
 
         nox -s unittest
 
-- Regenerate the python gRPC stubs
-
-        git clone https://github.com/googleapis/googleapis
-        nox -s protoc
-
 - Create a new `google_assistant_sdk` release in `dist`
 
         nox -s release
