+++ google-assistant-grpc/MANIFEST.in
