+++ google-assistant-grpc/google/assistant/embedded/v1alpha1/embedded_assistant_pb2_grpc.py
