+++ google-assistant-grpc/nox.py
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 import nox
-import os.path
-import tempfile
 
 
 @nox.session
@@ -22,35 +20,9 @@ def lint(session):
     session.interpreter = 'python3.4'
     session.install('pip', 'setuptools')
     session.install('docutils', 'flake8')
-    session.run('flake8',
-                'googlesamples', 'tests',
-                'nox.py', 'setup.py')
+    session.run('flake8', 'nox.py', 'setup.py')
     session.run('python', 'setup.py', 'check',
                 '--restructuredtext', '--strict')
-
-
-@nox.session
-@nox.parametrize('python_version', ['2.7', '3.4'])
-def unittest(session, python_version):
-    session.interpreter = 'python' + python_version
-    session.install('pip', 'setuptools')
-    session.install('pytest')
-    session.install('-e', '.[samples]')
-    session.run('py.test', 'tests')
-
-
-@nox.session
-@nox.parametrize('python_version', ['2.7', '3.4'])
-def endtoend_test(session, python_version):
-    session.interpreter = 'python' + python_version
-    session.install('pip', 'setuptools')
-    session.install('-e', '.[samples]')
-    temp_dir = tempfile.mkdtemp()
-    audio_out_file = os.path.join(temp_dir, 'out.raw')
-    session.run('python', '-m', 'googlesamples.assistant.grpc.pushtotalk',
-                '-i', 'tests/data/whattimeisit.riff',
-                '-o', audio_out_file)
-    session.run('test', '-s', audio_out_file)
 
 
 @nox.session
@@ -60,7 +32,7 @@ def protoc(session):
     session.run('python', '-m', 'grpc_tools.protoc',
                 '--proto_path=googleapis',
                 '--python_out=.',
-                '--grpc_python_out=.',
+                '--grpc_python_out=sdk/grpc/',
                 'googleapis/google/assistant/embedded/v1alpha1/'
                 'embedded_assistant.proto')
