+++ googlesamples/assistant/grpc/assistant_helpers.py
