+++ googlesamples/assistant/grpc/common_settings.py
