+++ googlesamples/assistant/grpc/requirements.txt
@@ -1,5 +1,5 @@
 google-assistant-grpc==0.0.1
-google-auth-oauthlib==0.0.1
+google-auth-oauthlib==0.1.0
 urllib3[secure]==1.20
 sounddevice==0.3.7
 click==6.7
