+++ tests/test_audio_helpers.py
